==> 1 Phrases Generation.ipynb <==
import pandas as pd
from tqdm import tqdm
import time

import numpy as np
np.random.seed(3)

import joblib
import plotly.express as px

import warnings
warnings.simplefilter('ignore')

import re
tqdm.pandas()

import nltk
from nltk.tokenize import word_tokenize
from nltk.stem.porter import PorterStemmer
from nltk.stem import SnowballStemmer

import spacy

from gensim import models
from tqdm import tqdm
import numpy as np

import plotly.graph_objects as go

from sklearn.feature_extraction.text import CountVectorizer

from gensim import corpora
from p_tqdm import p_map

df=pd.read_excel('/home/ubuntu/MyNotebooks/Data/wipo_ai_processed2.xlsx')

df.head()

nlp = spacy.load('en_core_web_md') #initializing stopwords
stop_words = list(spacy.lang.en.stop_words.STOP_WORDS) +['present','disclosure','method', 'system', 'use', 'useful', 'overall', 'have', 'having', 'said', 'relates', 'embodiment','methods','invention','consists','disclosed','inventions','apparatus','include','includes','provide','consist','comprise','comprises','includes','including','provided']

def process_text(text, stem=False): #clean text
    
    text = re.sub('[^A-Za-z]', ' ', text.lower()) #remove non-alphabets
    
    clean_text = [
         str(word.lemma_) for word in nlp(text)
         if word.pos_ in ['PROPN','NOUN','ADJ'] and str(word) not in stop_words
    ]
    clean_text=[x for x in clean_text if len(x)>1]
    return ' '.join(clean_text)

clean_text=p_map(process_text, df['Full Text'].astype(str).tolist())

from gensim.models.phrases import Phrases, Phraser

text_tokens=[x.split() for x in clean_text]

phrases=Phrases(text_tokens, min_count=5)

tokens=phrases[text_tokens]
for token in tokens[100:105]:
    print(token)

df['clean_text']=clean_text
df['phrases']=tokens
df.to_excel('wipo_processed_phrases.xlsx', index=False)
