==> 2 Topic Selection.ipynb <==
import pandas as pd
import joblib
from gensim. models import LdaMulticore, CoherenceModel

df=pd.read_excel('wipo_processed_phrases.xlsx')
df.head()

import ast
from gensim import corpora
from tqdm import tqdm
import numpy as np

np.random.seed(3)

tokens=[ast.literal_eval(x) for x in df.drop_duplicates('Full Text')['phrases'].tolist()]

id2w = corpora.Dictionary(tokens)
corpus=[]
for rev in tqdm(tokens):
    corpus.append(id2w.doc2bow(rev))
len(tokens)

import time
model={}
coherence_model_lda=[]
for t in tqdm(range(6,15)):
    model[t] = LdaMulticore(corpus, num_topics=t, id2word=id2w, passes=50,workers=15)
    joblib.dump(model[t], f'model_{t}.sav')
    vis[t] = pyLDAvis.gensim.prepare(model[t], corpus, dictionary=id2w)
    coherence_model_lda.append(CoherenceModel(model=model[t], texts=tokens, dictionary=id2w, coherence='c_v').get_coherence())
    if t%2==0:
        time.sleep(5)

import plotly.graph_objects as go
fig = go.Figure(data=go.Scatter(x=list(range(16,20)), y=coherence_model_lda, mode='lines+markers'))
fig.update_layout(template='simple_white')
fig.show()
