==> 3 Topic labeling.ipynb <==
import pandas as pd
from tqdm import tqdm
import time

import numpy as np
np.random.seed(3)

import plotly.express as px

import warnings
warnings.simplefilter('ignore')

import re
tqdm.pandas()

import nltk
from nltk.tokenize import word_tokenize
from nltk.stem.porter import PorterStemmer
from nltk.stem import SnowballStemmer

import spacy

from sklearn.feature_extraction.text import CountVectorizer

from gensim import corpora
from p_tqdm import p_map
import ast

import joblib
model=joblib.load('model_7.sav')
df=pd.read_excel('wipo_processed_phrases.xlsx')
clean_text=df['clean_text'].tolist()
tokens=[ast.literal_eval(x) for x in df['phrases'].tolist()]

clean_text=df[~(df['clean_text'].isnull())]['clean_text'].tolist()

topics = model.print_topics(num_words=30)
for topic in topics:
    print(topic[0], re.sub('[0-9.*]', '',topic[1]))
    print()

labels={
    0: 'Pharmaceutical Drug Development',
    1: 'Networks/IOT',
    2: 'Chemical Manufacturing',
    3: 'Electronics and Semiconductor Manufacturing',
    4: 'Biomedical Informatics',
    5: 'Photography and Imaging',
    6: 'Digital Data Processing'
}
